Disable auto-refresh when switching pages
